+++ backend/src/main/java/com/dongnemarket/product/entity/Product.java
@@ -15,6 +15,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -40,7 +41,7 @@ public class Product extends BaseTimeEntity {
 	private String description;
 
 	@Column(nullable = false)
-	private Integer price;
+	private BigDecimal price;
 
 	@Enumerated(EnumType.STRING)
 	@Column(nullable = false, length = 20)
@@ -62,7 +63,7 @@ public class Product extends BaseTimeEntity {
 	}
 
 	private Product(Member member, Category category, String title, String description,
-					Integer price, String region) {
+					BigDecimal price, String region) {
 		this.member = member;
 		this.category = category;
 		this.title = title;
@@ -75,7 +76,7 @@ public class Product extends BaseTimeEntity {
 	}
 
 	public static Product create(Member member, Category category, String title, String description,
-								 Integer price, String region) {
+								 BigDecimal price, String region) {
 		return new Product(member, category, title, description, price, region);
 	}
 
@@ -91,7 +92,7 @@ public class Product extends BaseTimeEntity {
 		this.viewCount++;
 	}
 
-	public void update(Category category, String title, String description, Integer price, String region) {
+	public void update(Category category, String title, String description, BigDecimal price, String region) {
 		this.category = category;
 		this.title = title;
 		this.description = description;
@@ -135,7 +136,7 @@ public class Product extends BaseTimeEntity {
 		return description;
 	}
 
-	public Integer getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
@@ -158,4 +159,5 @@ public class Product extends BaseTimeEntity {
 	public LocalDateTime getDeletedAt() {
 		return deletedAt;
 	}
+
 }
